day6-user

+++ internal/model/history.go
@@ -9,9 +9,9 @@ import (
 // History(Audit_Trail) tracking operation by admin depends on
 type History struct {
 	gorm.Model
-	Id        uint      `json:"id"`
-	UserId    uint      `json:"userId"`
-	OrderId   uint      `json:"orderId"`
+	Id        int64     `json:"id"`
+	UserId    int64     `json:"userId"`
+	OrderId   int64     `json:"orderId"`
 	Operation int       `json:"operation"` //Enums Operation
 	CreatedAt time.Time `json:"createdAt"`
 	UpdatedAt time.Time `json:"updatedAt"`
